LayoutTests/platform/ios-simulator/fast/css/draggable-region-parser-expected.txt

This tests the parser and serializer for draggable region values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



FAIL parseDraggableRegion('drag') should be drag (of type string). Was null (of type object).
FAIL parseDraggableRegion('no-drag') should be no-drag (of type string). Was null (of type object).

Tests for variations in whitespace that are allowed.

FAIL parseDraggableRegion('  drag  ') should be drag (of type string). Was null (of type object).
FAIL parseDraggableRegion('  no-drag    ') should be no-drag (of type string). Was null (of type object).

Tests for case insensitivity.

FAIL parseDraggableRegion('Drag') should be drag (of type string). Was null (of type object).
FAIL parseDraggableRegion('DrAG') should be drag (of type string). Was null (of type object).
FAIL parseDraggableRegion('NO-DrAG') should be no-drag (of type string). Was null (of type object).
FAIL parseDraggableRegion('NO-DRAG') should be no-drag (of type string). Was null (of type object).

Some expected failures.

PASS parseDraggableRegion('') is null
PASS parseDraggableRegion(' ') is null
PASS parseDraggableRegion('x') is null
PASS parseDraggableRegion('x-') is null
PASS parseDraggableRegion('x()') is null
PASS parseDraggableRegion('x(a)') is null
PASS parseDraggableRegion('none x') is null
PASS parseDraggableRegion('none x()') is null
PASS parseDraggableRegion('none x(a)') is null

PASS successfullyParsed is true

TEST COMPLETE
